Added single OWASP ASVS file

+++ get_dataset.sh
@@ -4,26 +4,41 @@ mkdir RawDatasets
 cd RawDatasets
 
 # SecReq Dataset
-wget -O SecReq.zip https://www.dropbox.com/sh/mcvx5ium0zx7bly/AABfJaFt0nWvjiNJs1RUYf_Pa?dl=1
-unzip SecReq.zip -d SecReq
-rm SecReq.zip
+if [ ! -d SecReq ]; then
+    wget -O SecReq.zip https://www.dropbox.com/sh/mcvx5ium0zx7bly/AABfJaFt0nWvjiNJs1RUYf_Pa?dl=1
+    unzip SecReq.zip -d SecReq
+    rm SecReq.zip
+fi
 
 # # PROMISE Dataset
-wget -O Promise.tar https://zenodo.org/record/268542/files/nfr.tar?download=1
-tar -xvf Promise.tar
-rm Promise.tar
-sed -i '45s/.*/@ATTRIBUTE class {F,A,L,LF,MN,O,PE,SC,SE,US,FT,PO}/' ./nfr/nfr.arff
+if [ ! -d nfr ]; then
+    wget -O Promise.tar https://zenodo.org/record/268542/files/nfr.tar?download=1
+    tar -xvf Promise.tar
+    rm Promise.tar
+    sed -i '45s/.*/@ATTRIBUTE class {F,A,L,LF,MN,O,PE,SC,SE,US,FT,PO}/' ./nfr/nfr.arff
+fi
 
 # Concordia Dataset
-wget -O Concordia.tar.gz https://www.semanticsoftware.info/system/files/NFRClassifier.tar.gz
-tar -xzf Concordia.tar.gz
-rm Concordia.tar.gz
+if [ ! -d NFRClassifier ]; then
+    wget -O Concordia.tar.gz https://www.semanticsoftware.info/system/files/NFRClassifier.tar.gz
+    tar -xzf Concordia.tar.gz
+    rm Concordia.tar.gz
+fi
 
 # CCHIT Dataset
-wget -O CCHIT.xls https://www.dropbox.com/s/7pe4xq0ntwbbrlx/CCHIT%20Certified%202011%20Ambulatory%20EHR%20Criteria%2020110517.xls?dl=1
+if [ ! -f CCHIT.xls ]; then
+    wget -O CCHIT.xls https://www.dropbox.com/s/7pe4xq0ntwbbrlx/CCHIT%20Certified%202011%20Ambulatory%20EHR%20Criteria%2020110517.xls?dl=1
+fi
+
+# OWASP Application Security Verification Standard
+if [ ! -f OWASP.csv ]; then
+    wget -O OWASP.csv https://raw.githubusercontent.com/OWASP/ASVS/v4.0.2/4.0/docs_en/OWASP%20Application%20Security%20Verification%20Standard%204.0.2-en.csv
+fi
 
 cd ..
-python3 -m venv env
+[ ! -d env ] && python3 -m venv env
 . ./env/bin/activate
 pip3 install pandas liac-arff xlrd
-python3 prepare_data.py --sec_req ./RawDatasets/SecReq --promise ./RawDatasets/nfr/nfr.arff --concord ./RawDatasets/NFRClassifier/gate/application-resources/Requirements/ --cchit ./RawDatasets/CCHIT.xls
+python3 prepare_data.py --sec_req ./RawDatasets/SecReq --promise ./RawDatasets/nfr/nfr.arff \
+    --concord ./RawDatasets/NFRClassifier/gate/application-resources/Requirements/ \
+    --cchit ./RawDatasets/CCHIT.xls --owasp ./RawDatasets/OWASP.csv
